\name{agrostis}
\alias{agrostis}
\docType{data}
\title{
Agrostis variabiis cover measurements
}
\description{
Percent cover of the grass \emph{Agrostis variablis} at 25 alpine snowbank sites in the Absaroka- Bearthooth Mountains.
}

\usage{data(agrostis)}
\format{
  A data frame with 25 observations on the following 2 variables.
  \describe{
    \item{site}{Site number}
    \item{cover}{Percent cover measure}
  }
}
\keyword{datasets}

