\name{bats}
\alias{bats}
\docType{data}
\title{Bat forearm length as a function of bat age}
\description{
Data from northern myotis bats (\emph{Myotis septentrionalis}) captured in the field in Vermillion County Indiana in 2000.   
}
\usage{data(bats)}
\format{
  The dataframe has 2 columns:
  \describe{
    \item{days}{The age of the bats in days.}
    \item{forearm.length}{The length of the forearm in millimeters.} 
        }
}
\source{
Krochmal, A. R., and D. W. Sparks (2007)  \emph{Journal of Mammalogy}. 88(3): 649-656.
}
\keyword{datasets}
