\name{bear}
\alias{bear}
\docType{data}
\title{Grizzly bear litter sizes

}
\description{
Counts of grizzly bear (\emph{Ursus arctos}) litter sizes from the Greater Yellowstone Ecosystem from 1973-2010.  
}
\usage{data(bear)}
\format{
  A data frame with 38 observations on the following 5 variables.
  \describe{
    \item{\code{Year}}{Year.}
    \item{\code{X1.cub}}{The number of litters with one cub.}
    \item{\code{X2.cub}}{The number of litters with two cubs.}
    \item{\code{X3.cub}}{The number of litters with three cubs.}
    \item{\code{X4.cub}}{The number of litters with four cubs.}
  }
}

\source{
Haroldson, M. A (2010)  Assessing trend and estimating population size from counts of 
	unduplicated females. Pgs 10-15 in Schwartz, C. C., Haroldson, M. A., and K. West editors.  
  \emph{Yellowstone grizzly bear investigations: annual report of the Interagency grizzly bear study team, 2010}.  U. S. Geological Survey, Bozeman, MT. 

}
\keyword{datasets}

