\name{beetle}
\alias{beetle}
\docType{data}
\title{
Wood boring beetle data.
}
\description{
Saint Germain et al. (2007) modeled the presence absence of a saprophytic wood boring beetle (\emph{Anthophylax attenuatus}) as a function of the wood density of twenty-four decaying aspen trees (\emph{Populus tremuloides}) in Western Quebec Canada.  
}
\usage{data(beetle)}
\format{
  A data frame with 24 observations on the following 4 variables.
  \describe{
    \item{\code{Snag}}{Snag identifier}
    \item{\code{Yrs.since.death}}{The number of years since death, determined using dendrochronological methods.}
    \item{\code{Wood.density}}{The density of the decaying wood (dry weight/volume) in units of g cm\eqn{-3}.}
    \item{\code{ANAT}}{Beetle presence/absence (1/0)}
  }
}
\source{
Saint-Germain, M., Drapeau, P., and C. Buddle  (2007)  Occurrence patterns of aspen-
feeding wood-borers (Coleoptera: Cerambycidae) along the wood decay
gradient: active selection for specific host types or neutral mechanisms? \emph{Ecological Entomology}  32: 712-721.
}
\keyword{datasets}

