\name{bombus}
\Rdversion{1.1}
\alias{bombus}
\docType{data}
\title{Bombus pollen data.
}
\description{To investigate how pollen removal varied with reproductive caste in bemblebees (Bombus sp.) Harder and Thompson (1989) recorded the proportion of pollen removed by thirty five bumblebee queens and twelve worker bees.
}
\usage{data(bombus)}
\format{
This data frame contains the following columns:
   \describe{
    \item{pollen}{
      A numeric vector indicating the proportion of pollen removed.
         }
    \item{caste}{
      A character string vector indicating whether a bee was a worker "W" or a queen "Q".
         }
    }
  }
\source{
Harder, L. D. and Thompson, J. D.  1989.  Evolutionary options for maximizing pollen dispersal in animal pollinated plants.  \emph{American Naturalist} 133: 323-344.
}
\keyword{datasets}
