\name{book.menu}
\alias{book.menu}
\title{Pulldown menus for 'asbio' interactive graphical functions
}
\description{
Provides a pulldown GUI menus to allow access to \pkg{asbio} statistical and biological graphical demos, e.g. \code{\link{anm.ci}}, \code{\link{samp.dist}}, \code{\link{anm.loglik}}, etc.  The function \code{demos} is deprecated.  The function \code{table.menu} currently provides links to over 100 distinct GUIs (many of these are slaves to other GUIs), which in turn provide distinct graphical demonstrations. The GUIs work best with an SDI R interface.   
}
\usage{
book.menu()
}


\author{Ken Aho and Roy Hill (Roy created the \code{book.menu} GUI)}

\seealso{
\code{\link{anm.coin}},  \code{\link{anm.ci}},  \code{\link{anm.die}},  \code{\link{anm.exp.growth}},  \code{\link{anm.geo.growth}},  \code{\link{anm.log.growth}}, \code{\link{anm.loglik}},  \code{\link{anm.LVcomp}},  \code{\link{anm.LVexp}},  \code{\link{anm.ls}},  \code{\link{anm.ls.reg}},  \code{\link{anm.transM}}, \code{\link{see.lmu.tck}}, \code{\link{samp.dist}}, \code{\link{see.HW}},  \code{\link{see.logic}},  \code{\link{see.M}},   \code{\link{see.nlm}}, \code{\link{see.norm.tck}},  \code{\link{see.power}},  \code{\link{see.regression}},   \code{\link{see.typeI_II}}, \code{\link{selftest.se.tck1}},  \code{\link{shade.norm}}, \code{\link{Venn}}} 
\examples{
\dontrun{
book.menu()
}
}
\keyword{graphs}

