\name{bromus}
\alias{bromus}
\docType{data}
\title{Bromus tectorum dataset}
\description{
Cheatgrass (\emph{Bromus tectorum}) is an introduced annual graminoid that has invaded vast areas of sagebrush steppe in the intermountain west. Because it completes its vegetative growth stage relatively early in the summer, it leaves behind senescent biomass that burns easily.  As a result areas with cheatgrass often experience a greater frequency of summer fires. A number of dominant shrub species in sagebrush steppe are poorly adapted to fire.   As a result, frequent fires can change a community formerly dominated by shrubs to one dominated by cheatgrass.  Nitrogen can also have a strong net positive effect on the cheatgrass biomass.  A study was conducted at the Barton Road Long Term Experimental Research site (LTER) in Pocatello Idaho to simultaneously examine the effect of shrub removal and nitrogen addition on graminoid productivity.   
}
\usage{data(bromus)}
\format{
  The dataframe has 3 columns:
  \describe{
    \item{Plot}{Plot number.}
    \item{Biomass}{Grass biomass in grams per meter squared.} 
    \item{Trt}{Treatment.  C = Control, LN = Low nitrogen, HN = Hi Nitrogen, SR = Shrub removal.}
    }
}
\keyword{datasets}
