\name{cliff.sp}
\alias{cliff.sp}
\docType{data}
\title{Yellowstone NP cliff community data}
\description{
A subset of a dataset collected by Aho (2006) which describes  the distribution of communities of lichens and vascular and avascular plant species on montane cliffs in Northeast Yellowstone National Park.  Of particular interest was whether substrate (limestone or andesitic conglomerate) or water supply influenced community composition. 
}
\usage{data(cliff.sp)}
\details{
Responses are average counts from two 10 x 10 point frames at 54 sites. Abundance data are for eleven species, 9 lichens, 3 mosses, and 2 vascular plants. Data were gathered in the summer of 2004 on two andesitic/volcanic peaks (Barronette and Abiathar) with sedimentary layers at lower elevations.  
}
\source{Aho, K.(2006)  \emph{Alpine Ecology and Subalpine Cliff Ecology in the Northern Rocky 
Mountains}.  Doctoral dissertation, Montana State University, 458 pgs.}
\keyword{datasets}
