\name{crab.weight}
\alias{crab.weight}
\docType{data}
\title{
crab gill and body weight data
}
\description{
Gill weight and body weight data for 12 striped shore crabs (\emph{Pachygrapsus crassipes}).
}
\usage{data(crab.weight)}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{gill.wt}}{Gill weight in mg}
    \item{\code{body.wt}}{Body weight in grams}
  }
}
\source{
Sokal, R. R., and F. J. Rohlf  (2012)  \emph{Biometry, 4th edition}.  W. H. Freeman and Co., New York. 
}
\keyword{datasets}

