\name{crabs}
\alias{crabs}
\docType{data}
\title{
Agresti crabs dataset
}
\description{
Horseshoe crab satellite counts as a function of crab phenotype.
}
\usage{data(crabs)}
\format{
  A data frame with 173 observations on the following 5 variables.
  \describe{
    \item{\code{color}}{A factor with levels \code{1} = light medium, \code{2} = medium, \code{3} = dark medium, \code{4} = dark.}
    \item{\code{spine}}{A factor with levels \code{1} = both good, \code{2} = one worn or broken, \code{3} = both worn or broken.}
    \item{\code{width}}{Crab carapace width in cm.}
    \item{\code{satell}}{Number of satellites.}
    \item{\code{weight}}{Crab weight in in kg.}
  }
}
\source{
Agresti, A.  (2012)  \emph{An Introduction to Categorical Data Analysis, 3rd edition}.  Wiley, New Jersey.
}
\references{
Brockman, H. J.  (1996) Satellite male groups in horseshoe crabs, \emph{Limulus polyphemus}.  \emph{Ethology} 102(1) 1-21.
}
\keyword{datasets}

