\name{cuckoo}
\alias{cuckoo}
\docType{data}
\title{
Tippett cuckoo egg data}
\description{
Part of a dataset detailing cuckoo egg lengths found in other birds nests.  Units are millimeters
}
\usage{data(cuckoo)}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{TP}}{Tree pipit}
    \item{\code{HS}}{Hedge sparrow}
    \item{\code{RB}}{Robin}
  }
}
\source{
Tippett, L. H. C. (1952) \emph{The Methods of Statistics, 4th Edition}.  Wiley.
}
\keyword{datasets}

