\name{deer}
\alias{deer}
\docType{data}
\title{Maternal deer data }
\description{
Monteith et al. (2009) examined the maternal life history characteristics of white-tailed deer (Odocoileus virginianus) originating from the Black Hills in southwestern South Dakota and from eastern South Dakota.  Because litter size and dam size affects offspring weight the investigators used proportional birth mass (dam mass/total litter mass) as a measure of reproductive investment by deer.  
}
\usage{data(deer)}
\format{
The dataframe contains 6 columns
\describe{
\item{Birth.Yr}{Year of birth.}   
\item{Litter.size}{Number of offspring.}         
\item{Region}{Categorical variable with two factor levels.  BH = Black Hills, ER = Eastern Region.}        
\item{Dam.weight}{Dam weight in kg.}
\item{Total.birth.mass}{Mass of litter in kg.}
\item{Prop.mass}{Total birth mass divided by dam weight.}
  }
}
\source{Monteith, K. L, Schmitz, L. E., Jenks, J. A., Delger, J. A., and R. T. Bowyer.  2009.  Growth of male white tailed deer: consequences of maternal effects.  \emph{Journal of Mammalogy} 90(3): 651-660.}
\keyword{datasets}
