\name{demos}
\alias{demos}
\title{Pulldown menu for 'asbio' interactive graphical functions.
}
\description{
Provides a pulldown GUI menu to access \pkg{asbio} graphical functions, e.g. \code{\link{anm.ci}}, \code{\link{samp.dist}}, \code{\link{loglik.plot}}, etc.  This function is early in its development.  
}
\usage{

demos()
}


\author{Ken Aho

}
\seealso{
\code{\link{anm.ci}}, \code{\link{samp.dist}}, \code{\link{loglik.plot}}
}
\examples{
\dontrun{
demos()
}
}
\keyword{graphs}

