\name{dO2}
\alias{dO2}
\docType{data}
\title{
Dissolved levels in locations above and below a town 
}
\description{
Dissolved O\eqn{_2} readings in ppm for 15 random locations above and below a riverside community. 
}
\usage{data(dO2)}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{\code{O2}}{Dissolved O\eqn{_2} levels in ppm.}
    \item{\code{location}}{River flow location with respect to town.  Levels are \code{Above} \code{Below}.}
  }
}

\source{
Ott, R. L., and M. T. Longnecker  (2004)  \emph{A First Course in Statistical Methods}.  Thompson. 
}
\keyword{datasets}

