\name{drugs}
\alias{drugs}
\docType{data}
\title{
Contingency data for high school marijuana, alcohol, and cigarette use
}
\description{
Agresti (2012) incldued a three way contingency table describing cigarette, alcohol, and marijuana use of high school students in Dayton Ohio.
}
\usage{data(drugs)}
\format{
  A data frame with 8 observations on the following 4 variables.
  \describe{
    \item{\code{alc}}{Alcohol use. A factor with levels \code{N} \code{Y}.}
    \item{\code{cig}}{Cigarette use.  A factor with levels \code{N} \code{Y}.}
    \item{\code{mari}}{Marijuana use. A factor with levels \code{N} \code{Y}}
    \item{\code{count}}{Counts for the cross-classification.}
  }
}

\source{
Agresti, A.  2012.  Categorical data analysis, 3rd edition.  New York.  Wiley. 
}
\keyword{datasets}

