\name{Fbird}
\alias{Fbird}
\docType{data}
\title{Frigatebird drumming frequency data}
\description{
Male magnificent frigatebirds (Fregata magnificens) have an enlarged red throat pouch that has probably evolved as the result of sexual selection.  
During courtship displays males attract females by displaying this pouch and using it to make a drumming sound. Madsen et al. (2004) noted that conditions 
(e.g. oblique viewing angles) often limit females' ability to appraise pouch size exactly.  Since females choose mates based on pouch size, 
a question of interest is whether females could use the pitch of the pouch drumming as an indicator of pouch size. 
Madsen et al. (2004) estimated the pouch volume and fundamental drumming frequency for forty males at Isla Isabel in Nayarit Mexico.  
Eighteen of these observations are in this dataset.
}
\usage{data(Fbird)}
\format{
The dataframe contains two variables: 
  \describe{
    \item{vol}{
     Pouch volume in (cm^3). 
    }
    \item{freq}{
     Frequency of drumming (in Hz)
    }
    }
}
\source{Madsen, V., Balsby, T.J.S., Dabelsteen, T., and J.L. Osorno (2004) Bimodal signaling of 
a sexually selected trait: gular pouch drumming in the magnificent frigatebird. \emph{Condor} 106: 156-160.
} 
\keyword{datasets}
