\name{fire}
\alias{fire}
\docType{data}
\title{
Fire data from Yellowstone National Park
}
\description{
Fires from 1988 constituted the largest conflagration in the history of Yellowstone National Park.  This dataframe lists burned areas for ten Yellowstone stream catchments (Robinson et al. 1994).   
}
\usage{data(fire)}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{fire}}{Burn area in, in \deqn{hectares^2}.}
    \item{\code{stream}}{A factor with levels \code{Blacktail} \code{Cache} \code{EF.Blacktail} \code{Fairy} \code{Hellroaring} \code{Iron.Springs} \code{Pebble} \code{Rose} \code{SF.Cache} \code{Twin}}
  }
}

\source{
Robinson, C. T., Minshall, G. W., and S. R. Rushforth  (1994)  The effects of the 1988 
wildfires on diatom assemblages in the streams of Yellowstone National Park.  Technical Report NPS/NRYELL/NRTR-93/XX. 
}

\keyword{datasets}

