\name{fly.sex}
\alias{fly.sex}
\docType{data}
\title{
Fly sex and longevity
}
\description{
Partridge and Farqaur (1981) studied the effect of the number of mating partners on the longevity of fruit flies.  Five different mating treatments were applied to single male fruit flies.  As a concomitant variable thorax length was measured.
}
\usage{data(fly.sex)}


\format{
  A data frame with 125 observations on the following 3 variables.
  \describe{
    \item{\code{treatment}}{a factor with levels \code{1} = One virgin female per day, \code{2} = eight virgin females per day, \code{3} = a control group with one newly inseminated female per day, \code{4} = a control group with eight newly inseminated females per day, 
    and \code{5} a control group with no added females.}
    \item{\code{longevity}}{Age in days.}
    \item{\code{thorax}}{Thorax length in mm.}
  }
}

\source{
Quinn, G. P., and M. J. Keough.  2002.  \emph{Experimental Design and Data Analysis for 
Biologists}.  Cambridge University Press. 
}
\references{
Partridge, L., and Farquhar, M. (1981), Sexual activity and the lifespan of male fruit flies, \emph{Nature} 294, 580-581.
}
\keyword{datasets}

