\name{frog}
\alias{frog}
\docType{data}
\title{Australian frog calls following fire
}
\description{
Driscoll and Roberts (1997) examined the impact of fire on the Walpole frog (\emph{Geocrinia lutea}) in catchments in Western Australia by counting the number of calling males in six paired burn and control sites for three years following spring burning in 1991. 
}
\usage{data(frog)}

\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{catchment}}{A factor with levels \code{angove} \code{logging} \code{newpipe} \code{newquinE} \code{newquinW} \code{oldquinE}.}
    \item{\code{frogs}}{The difference in the number of male frog calls for control - burned sites.}
    \item{\code{year}}{Year.}
  }
}
\source{
Quinn, G. P., and M. J. Keough (2002) \emph{Experimental Design and Data Analysis for Biologists}. Cambridge University Press.
}
\references{
Driscoll, D. A., and J. D. Roberts. 1997. Impact of fuel-reduction burning on the frog Geocrinia lutea in southwest Western Australia. \emph{Australian Journal of Ecology} 22:334-339.
}
\keyword{datasets}

