\name{fruit}
\alias{fruit}
\docType{data}
\title{
Fruit weight data from Littell et al. (2002)
}
\description{
Valencia orange tree fruit weights are measured at harvest with respect to five irrigation treatment applied in eight blocks in a RCBD.  
}
\usage{data(fruit)}
\format{
  A data frame with 40 observations on the following 3 variables.
  \describe{
    \item{\code{block}}{a factor describing eight blocks}
    \item{\code{irrig}}{a factor with levels \code{basin} \code{flood} \code{spray} \code{sprinkler} \code{trickle}}
    \item{\code{fruitwt}}{a numeric vector}
  }
}
\source{
Littell, R. C., Stroup, W. W., and R. J. Fruend (2002)  \emph{SAS for linear models}.  John 
Wiley and Associates.
}
\keyword{datasets}
