\name{G.mean}
\alias{G.mean}
\title{Geometric mean}
\description{Calculates the geometric mean.}
\usage{G.mean(x)}
\arguments{
  \item{x}{A vector of quantitative data.}
}
\value{Returns the geometric mean.}
\author{Ken Aho}
\seealso{\code{\link{H.mean}}, \code{\link{HL.mean}}}
\examples{
x<-c(2,1,4,5,6,2.4,7,2.2,.002,15,17,.001)
G.mean(x)
}

