\name{goats}
\alias{goats}
\docType{data}
\title{Mountain goat data from Yellowstone National Park}
\description{
Mount goat (\emph{Oreomnos americanus}) feces data and soil nutrient data for eight different mountains in the Northern Absarokas in Yellowstone National Park.    
}
\usage{data(goats)}
\format{
  The dataframe has 3 columns:
  \describe{
    \item{\code{feces}}{feces concentration (Percent occurrence per 0.1, m^2 plot).}
    \item{\code{NO3}}{Nitrate concentration in ppm.} 
    \item{\code{organic.matter}}{Organic matter concentration (LOI) as a percentage.} 
        }
}
\source{
Aho, K. (2012)  \emph{Management of introduced mountain goats in Yellowstone National Park (vegetation analysis along a mountain goat gradient).  PMIS:  105289}.  Report prepared for USDA National Park Service. 150 pp.
}
\keyword{datasets}
