\name{H.mean}
\alias{H.mean}
\title{Harmonic mean}
\description{Calculates the harmonic mean.}
\usage{H.mean(x)}
\arguments{
  \item{x}{Vector of quantitative data.}
}
\value{
Returns the harmonic mean.
}
\seealso{\code{\link{G.mean}}, \code{\link{HL.mean}}}
\author{Ken Aho}
\examples{
x<-c(2,1,4,5,6,2.4,7,2.2,.002,15,17,.001)
H.mean(x)
}

