\name{HL.mean}
\alias{HL.mean}
\title{Hodges-Lehman estimator of location}
\description{Calculates Hodges-Lehman estimates of location.}
\usage{HL.mean(x)}
\arguments{
  \item{x}{A vector of quantitative data.}
}
\author{Ken Aho}
\seealso{\code{\link{H.mean}}, \code{\link{G.mean}}}
\examples{x<-c(2,1,4,5,6,2.4,7,2.2,.002,15,17,.001)
HL.mean(x)
}


