\name{illusions}
\alias{illusions}
\title{
Visual illusions illustrating human perception errors.
}
\description{
In development, currently displays three illusions.  Illusion 3 is from Yihui Xie's package \pkg{animation}.
}
\usage{

illusions(ill.no = 1)
}

\arguments{

  \item{ill.no}{Integer in 1:3 describing which illusion number to view.

}
}

\author{
Ken Aho.  Illusion 3 uses code from Yihui Xie's package animation.
}
\examples{
illusions(1)
illusions(2)
illusions(3)
}
\keyword{graphs}
