\name{ipomopsis}
\alias{ipomopsis}
\docType{data}
\title{
Ipomopsis fruit yield data
}
\description{
The following question is based on data from Crawley (2007).  We are interested in the effect of grazing on seed production in the plant scarlet gilia \emph{Ipomopsis aggregata} .  Forty plants were allocated to two treatments, grazed and ungrazed.  Grazed plants were exposed to rabbits during the first two weeks of stem elongation.  They were then protected from subsequent grazing by the erection of a fence and allowed to continue growth. Because initial plant size may influence subsequent fruit production, the diameter of the top of the rootstock was measured before the experiment began.  At the end of the experiment, fruit production (dry weight in milligrams) was recorded for each of the forty plants.  
}
\usage{data(ipomopsis)}
\format{
  A data frame with 40 observations on the following 3 variables.
  \describe{
    \item{\code{root}}{Rootstock diameter in mm}
    \item{\code{fruit}}{Fruit dry weight in mg}
    \item{\code{grazing}}{a factor with levels \code{Grazed} \code{Ungrazed}}
  }
}

\source{
Website associated with --
Crawley, M. J.  2007.  \emph{The R book}.  Wiley. 
}
\keyword{datasets}

