\name{K}
\alias{K}
\docType{data}
\title{
Soil potassium analyses from 8 laboratories
}
\description{
Jacobsen et al. (2002) sent nine "identical" soil samples to eight soil testing 
laboratories in the Great Plains region of the Central United States over a three year period of time. 
 Among other characteristics the labs were paid to measure soil potassium.  
A question of interest was whether the labs would produce identical analytical results.
}
\usage{data(K)}
\format{
  A data frame with 72 observations on the following 2 variables.
  \describe{
    \item{\code{K}}{Soil K in mg/kg}
    \item{\code{lab}}{Laboratories, a factor with levels \code{B} \code{D} \code{E} \code{F} \code{G} \code{H} \code{I} \code{J}}
  }
}
\source{
Jacobsen, J. S., Lorber, S. H., Schaff, B. E., and C. A. Jones (2002)  Variation in soil 
fertility test results from selected Northern Great Plains laboratories.  \emph{Commun. Soil Sci. plant Anal.}, 33(3&4): 303-319.
}

\keyword{datasets}

