\name{larrea}
\alias{larrea}
\docType{data}
\title{
Creosote bush counts 
}
\description{
Phillips and MacMahon (1981) conducted an extensive study of \emph{Larrea tridentata} (creosote bush) distributions in the Mojave 
and Sonoran deserts from several life stage classes based areal coverage: Life stage 1 (102 -103 cm\eqn{^2})	Life stage 2 (103 -104 cm\eqn{^2}), and	Life stage 3
(104 -105 cm\eqn{^2}).  Data were generated to approximate the results of the authors.
}
\usage{data(larrea)}
\format{
  A data frame with 25 observations on the following 3 variables.
  \describe{
    \item{\code{class1}}{Counts from life stage 1}
    \item{\code{class2}}{Counts from life stage 2}
    \item{\code{class3}}{Counts from life stage 3}
  }
}

\references{
Phillips, D. L., and J. A. MacMahon  (1981)  Competition and spacing patterns in desert 
shrubs.  \emph{Journal of Ecology} 69(1): 97-115. 

}
\keyword{datasets}

