\name{lmr.tck}
\alias{lmr.tck}
\alias{pm1}
\title{
Regression linear model derivation from linear algebra
}
\description{
Given Y and X matrices a regression linear model is demonstrated using matrix algebra.
}
\usage{
lmr.tck()
pm1(Y, X, sz=1)
}

\arguments{

  \item{Y}{
Response variable
}
  \item{X}{
Explanatory variables
}
  \item{sz}{
Text expansion factor
}
}
\details{
Newly implemented, X requires a Y intercept variable and at least one other variable.
}
\author{Ken Aho
}
\seealso{\code{\link{lm}}}
\keyword{graphs}
