\name{magnets}
\alias{magnets}
\docType{data}
\title{Magnet pain relief data}
\description{
Magnets have long been used as an alternative medicine, particularly in the Far East, for speeding the recovery of broken bones and to aid in pain relief.  
Valbona et al. (1997) tested whether chronic pain experienced by post-polio patients could be treated with magnetic fields applied directly to pain trigger points.  
The investigators identified fifty subjects who not only had post-polio syndrome, but who also experienced muscular or arthritic pain.  
Magnets were applied to pain trigger points in 29 randomly selected subjects, and in the other 21 a placebo was applied.  
The patients were asked to subjectively rate pain on a scale from one to ten before and after application of the magnet or placebo.

}
\usage{data(magnets)}
\format{
The dataframe contains 4 columns
\describe{
\item{Score_1}{Reported pain level before application of treatment.}   
\item{Score_2}{Reported pain level after application of treatment.}         
\item{Active}{Categorical variable indicating whether the device applied was active (magent) or inactive (placebo).}        
  }
}
\source{Vallbona, C. et al  (1997)  Response of pain to static magnetic fields in postpolio patients, 
a double blind pilot study. \emph{Archives of Physical Medicine and Rehabilitation}.  78: 1200-1203.
}
\keyword{datasets}
