\name{montane.island}
\alias{montane.island}
\docType{data}
\title{
Mountain island biogeographic data
}
\description{
Lomolino et al. (1989) investigated the relationship between the area of montane  forest patches (islands) and the richness of mammal fauna in the Southwestern United States.  This dataset contains richness and area information for 27 montane islands. 
}
\usage{data(montane.island)}
\format{
  A data frame with 27 observations on the following 3 variables.
  \describe{
    \item{\code{Island.name}}{a factor with levels \code{Abajos} \code{Animas} \code{Blacks} \code{Capitans} \code{Catalinas} \code{Chirucahuas} \code{Chuskas} \code{Guadalupes} \code{Huachucas} \code{Hualapais} \code{Lasals} \code{Magdalenas} \code{Manzanos} \code{Mogollon} \code{Mt. Taylor} \code{N. Rincon} \code{N. Uncompaghre} \code{Navajo} \code{Organs} \code{Pinalenos} \code{Prescotts} \code{S. Kabib} \code{Sacramentos} \code{San Mateos} \code{Sandias} \code{Santa Ritas} \code{Zunis}.}
    \item{\code{Richness}}{A numeric vector; the number of species.}
    \item{\code{Area}}{A numeric vector; area in km\eqn{^2}.}
  }
}
\source{
Lomolino, M. V., Brown, J. H., and R. Davis (1989)  Island biogeography of montane 
forest mammals in the American Southwest.  \emph{Ecology} 70: 180-194. 
}
\keyword{datasets}

