\name{myeloma}
\alias{myeloma}
\docType{data}
\title{
Patient responses to myeloma drug treatments
}
\description{
Murakami et al. (1997) studied the effect of drugs treatments on levels of serum beta-2 microglobulin in patients with multiple myeloma.   
Serum beta-2 microglobulin is produced in the body as a result of myelomas, and thus can be used as an indicator of the severity of disease
}
\usage{data(myeloma)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{mglobulin}}{Levels of serum beta-2 microglobulin in mg/l}
    \item{\code{drug}}{Drug treatment strategy.  \code{Control} = sumerifon alone, \code{Trt} = malphalan and sumerifon.}
  }
}
\source{
Ott, R. L., and M. T. Longnecker  (2004)  \emph{A First Course in Statistical Methods}.  Thompson. 
}
\references{
Murakami, H., Ogawara, H., Morita, K.,  Saitoh, T., Matsushima, T., Tamura, J.,   Sawamura, M.,  Karasawa, M.,  Miyawaki, M.,, Schimano, S., Satoh, S., and  J Tsuchiy (1997)
Serum beta-2-microglobulin in patients with multiple myeloma treated with alpha interferon.  \emph{Journal of Medicine} 28(5-6):311-8. 
}
\keyword{datasets}

