\name{PCB}
\alias{PCB}
\docType{data}
\title{
PCBs and herring egg thickness
}
\description{
Thirteen sites in the Great Lakes were selected for a study to quantify PCB 
concentrations in 1982 and 1996 (Hughes et al. 1998).  At each site 9-13 American 
herring gull (Larus smithsonianus) eggs were randomly collected and tested for PCB content.
}
\usage{data(PCB)}
\format{
  A data frame with 26 observations on the following 3 variables.
  \describe{
    \item{\code{nest}}{Nest number}
    \item{\code{level}}{PCB levels microgram/gram of dry weight}
    \item{\code{year}}{a numeric vector}
}
}
\source{
Ott, R. L., and M. T. Longnecker  (2004)  \emph{A First Course in Statistical Methods}.  Thompson. 
}
\references{
Hughes, K. D., Weselogh, D. V., and B. M. Braune (1998) The ratio of DDE to PCB 
concentrations in Great Lakes herring gull eggs and its use in interpreting contaminants data. \emph{Journal of Great Lakes Research} 24(1): 12-31. 
}
\keyword{datasets}

