\name{pika}
\alias{pika}
\docType{data}
\title{
Nitrogen content of soils under pika haypiles.
}
\description{
Aho (1998) hypothesized that pikas worked as ecosystem engineers by building relatively rich soils (via decomposing haypiles and fecal accumulations) in otherwise barren scree.  
Soils from twenty one paired on-haypile and off-haypile sites were gathered from Rendezvous Peak Grand Teton National Park to determine if the habitats differed in total soil nitrogen.
}
\usage{data(pika)}
\format{
  A data frame with 22 observations on the following 2 variables.
  \describe{
    \item{\code{Haypile}}{a numeric vector}
    \item{\code{On.Off..N}}{a numeric vector}
  }
}
\references{
Aho, K., Huntly N., Moen J., and T. Oksanen (1998) Pikas (\emph{Ochotona princeps}: Lagomorpha) as allogenic engineers in an alpine ecosystem. \emph{Oecologia}.  114 (3): 405-409. 
}

\keyword{datasets}

