\name{PM2.5}
\alias{PM2.5}
\docType{data}
\title{
PM 2.5 pollutant data from Pocatello Idaho
}
\description{
PM 2.5 pollutants (those less than 2.5 microns in diameter) can be directly emitted from sources such as forest fires, or can form when gases discharged from power plants, industries and automobiles react in the air. Once inhaled, these particles can affect the heart and lungs and cause serious health problems. The DEQ began monitoring PM 2.5 pollutants in Pocatello Idaho in November 1998. 
}
\usage{data(PM2.5)}
\format{
  A data frame with 65 observations on the following 2 variables.
  \describe{
    \item{\code{Yr.mos}}{Year and month.  A factor with levels \code{1998 11} \code{1998 12} \code{1999 1} \code{1999 10} \code{1999 11} \code{1999 12} \code{1999 2} \code{1999 3} \code{1999 4} \code{1999 5} \code{1999 6} \code{1999 7} \code{1999 8} \code{1999 9} \code{2000 1} \code{2000 10} \code{2000 11} \code{2000 12} \code{2000 2} \code{2000 3} \code{2000 4} \code{2000 5} \code{2000 6} \code{2000 7} \code{2000 8} \code{2000 9} \code{2001 1} \code{2001 10} \code{2001 11} \code{2001 12} \code{2001 2} \code{2001 3} \code{2001 4} \code{2001 5} \code{2001 6} \code{2001 7} \code{2001 8} \code{2001 9} \code{2002 1} \code{2002 10} \code{2002 11} \code{2002 12} \code{2002 2} \code{2002 3} \code{2002 4} \code{2002 5} \code{2002 6} \code{2002 7} \code{2002 8} \code{2002 9} \code{2003 1} \code{2003 10} \code{2003 11} \code{2003 12} \code{2003 2} \code{2003 3} \code{2003 4} \code{2003 5} \code{2003 6} \code{2003 7} \code{2003 8} \code{2003 9} \code{2004 1} \code{2004 2} \code{2004 3}}
    \item{\code{PM2.5}}{A numeric vector describing PM 2.5 pollutant levels in \eqn{mu}g/\eqn{m^2}.}
  }
}
\source{
Idaho department of Environmental Quality
}
\keyword{datasets}
