\name{polyamine}
\alias{polyamine}
\docType{data}
\title{
Polyamine data from Hollander and  Wolfe (1999)
}
\description{
Polyamines are a class of organic compounds having two or more primary amino groups.  They appear to have a number of important functions including regulation of cell proliferation, cell differentiation, and cell death.  Polyamine plasma levels taken for healthy children of different ages were summarized by Hollander and Wolfe (1999).
}

\usage{data(polyamine)}
\format{
  A data frame with 25 observations on the following 2 variables.
  \describe{
    \item{\code{age}}{Child age in years (0 indicates newborn)}
    \item{\code{p.amine}}{Polyamine level in blood}
  }
}
\source{
Hollander, M., and  D. A. Wolfe  (1999) \emph{Nonparametric statistical methods}. New York: John Wiley & Sons. 
}
\keyword{datasets}

