\name{portneuf}
\alias{portneuf}
\docType{data}
\title{
Portneuf River longitudinal N and P data
}
\description{
Portneuf River data from the Siphon Road site near Pocatello Idaho, downstream from an elemental P refinery.
}
\usage{data(portneuf)}
\format{
  A data frame with 176 observations on the following 3 variables.
  \describe{
    \item{\code{date}}{Dates from 1998-01-15 to 2011-08-16}
    \item{\code{TKN}}{Total Kjeldahl nitrogen (measured as a percentage)}
    \item{\code{total.P}}{Total phosphorous (mg/L)}
  }
}
\source{
Idaho State Department of Environmental Quality
}
\keyword{datasets}
