\name{prostate}
\alias{prostate}
\docType{data}
\title{
Prostate cancer data
}
\description{Hastie et al. (2001) describe a cancer research study that attempted to associate prostate specific antigenand and a number of prognostic measures in mean with advanced prostate cancer.  
Data in the experiment were collected from 97 men who were about to undergo radial prostectomies. 
}
\usage{data(prostate)}
\format{
 A data frame with 97 observations on the following 4 variables.
  \describe{
    \item{\code{PSA}}{Serum prostate-specific albumin levle (mg/ml).}
    \item{\code{vol}}{Tumor volume (cc).}
    \item{\code{weight}}{Prostate weight (g).}
    \item{\code{Gleason}}{Pathologically determined grade of disease.  Summed scores were either 6, 7, or 8 with higher scores indicating worse prognosis.}
  }
}
\source{
Kutner, M. H., Nachtsheim, C. J., Neter, J., and W. Li  (2005)  \emph{Applied Linear Statistical Models, 5th edition}.  McGraw-Hill, Boston. 
}
\references{
Hastie, T., R. Tibshirani, and J. Friedman (2009) \emph{The Elements of Statistical Learning: Data Mining, Inference, and Prediction, Second Edition}. Springer.
}
\keyword{datasets}

