\name{refinery}
\alias{refinery}
\docType{data}
\title{Refinery CO dataset}
\description{
In the early 1990s an oil refinery northeast of San Francisco agreed with local air quality regulators [the Bay Area Air Quality Management District (BAAQMD)] to reduce carbon monoxide emissions.  Baselines for reductions were to be based on measurements of CO made by refinery personnel, and by independent measurements from BAAQMD scientists for the roughly the same time period 
}
\usage{data(refinery)}
\format{
The dataframe contains three columns:
  \describe{
\item{CO}{Carbon monoxide.  Measured in ppm.}
\item{Source}{Te source of measurements; either refinery or BAAQMD.}
\item{Date}{Month/Day/Year}
}
}
\source{
http://lib.stat.cmu.edu/DASL/Stories/MeasuringAirPollution.html; accessed 6/29/110
}
\keyword{datasets}
