\name{samp.dist.mech}
\alias{samp.dist.mech}
\alias{samp.dist.mech.tck}
\title{
Animated representation of sampling distribution basics
}
\description{
Mountain goats are randomly sampled 10 at a time and weighed [goat weights are normal \emph{N}(90.5, 225)], a mean weight is calculated from these measures and added to collection of mean weights in the form of a histogram. 
}
\usage{

samp.dist.mech(rep, int = 0.05)

samp.dist.mech.tck()

}
\arguments{

  \item{rep}{
Number of samples.  Should not greatly exceed 100.
}
  \item{int}{The time interval for animation (in seconds).  Smaller intervals speed up animation}
}

\author{
Ken Aho
}
\note{
Nice goat image from \url{http://all-free-download.com} 
}

\keyword{graphs}

