\name{savage}
\alias{savage}
\docType{data}
\title{
Mammalian BMR and biomass data from Savage et al. (2004)
}
\description{
Compilation of mammalian BMR and biomass data from the large
data sets used in the studies of Hart (1971),
Heusner (1991), Lovegrove (2000, 2003) and White &
Seymour (2003).  Data compiled by Savage (2004).
}
\usage{data("savage")}
\format{
  A data frame with 1006 observations on the following 9 variables.
  \describe{
    \item{\code{Order}}{Mammal order.}
	\item{\code{Family}}{Mammal family.}
	\item{\code{Species}}{Mammal genus and species.}
    \item{\code{Mass}}{Biomass in grams.}
    \item{\code{BMR}}{Basal metabolic rate in watts}
    \item{\code{AvgMass}}{Average mass, given multiple reports for a particular species.}
    \item{\code{AvgBMR}}{Average BMR, given multiple reports for a particular species.}
    \item{\code{References}}{Authorities from whom data were obtained.}
    \item{\code{Notes}}{Note concerning a repeated species name.}
  }
}
\source{
Savage, 548 V.M., Gillooly, J.F., Woodruff, W.H., West, G.B., Allen, A.P., Enquist, B.J., 
Brown, J.H. (2004) The predominance of quarter-power scaling in biology. \emph{Functional 
Ecology}, 18, 257-282.
}
\references{
Hart, J.S. (1971) \emph{Rodents in Comparative Physiology of Thermoregulation}, Vol. II Mammals (ed. G.C. Whittow),
pp. 2-149. Academic Press, New York.

Heusner, A.A. (1991) Size and power in mammals. \emph{Journal of Experimental Biology} 160, 25-54.

Lovegrove, B.G. (2000) The zoogeography of mammalian basal metabolic rate. \emph{American Naturalist} 156, 201-219.

Lovegrove, B.G. (2003) The influence of climate on the metabolic rate of small mammals: a slow-fast metabolic continuum.  \emph{Journal of Comparative Physiology B} 173, 87-112.

White, C.R. and Seymour, R.S. (2003) Mammalian basal metabolic rate is proportional to
584 body mass 2/3. \emph{Proceedings of the National Academy of Sciences}, 100, 4046-4049.
}
\keyword{datasets}

