\name{sc.twin}
\alias{sc.twin}
\docType{data}
\title{Matched pairs schizophrenia data}
\description{
Scientists have long been concerned with identifying physiological characteristics which result in a disposition for schizophrenia.  Early studies suggested that the volume of particular brain regions of schizophrenic patients may differ from non-afflicted individuals.  However these studies often contained confounding variables (e.g. socioeconomic status, genetics) which obscured brain volume/ schizophrenia relationships (Ramsey and Schafer 1997).  To control for confounding variables Suddath et al. (1990) examined 15 pairs of monozygotic twins where one twin was schizophrenic and the other was not.  Twins were located from an intensive search throughout the United States and Canada.  The authors used magnetic resonance imaging to measure brain volume of particular regions in the twin's brains. 
}
\usage{data(sc.twin)}
\format{
  The dataframe has 2 columns:
  \describe{
    \item{unaffected}{Left hippocampus volumes for unaffected twins.}
    \item{affected}{Left hippocampus volumes for affected twins.}
    }
  } 
\keyword{datasets}
