\name{see.ancova.tck}
\alias{see.ancova.tck}

\title{
Visualize ANCOVA mechanics
}
\description{
An interactive GUI to view ANCOVA meachnics.  Exp. power tries to simulate explanatory power in the concomitant variable.  It simply results in (1 - Exp. power) \eqn{\times} Residual SE.
}
\usage{

see.ancova.tck()
}
\author{
Ken Aho
}
\keyword{graphs}