\name{see.anova.tck}
\alias{see.anova.tck}

\title{
Interactive depiction of the ANOVA mechanism
}
\description{
Slider control of the means and (constant) variablility of three factor level populations.  AN ANOVA is run by randomly sampling these populations.
}
\usage{

see.anova.tck()
}


\author{
Ken Aho
}
\keyword{grpahs}
