\name{see.cor.range.tck}
\alias{see.cor.range.tck}
\title{
Depict the effect of range on correlation
}
\description{
Function interactively depicts the effect of the data range on association measures.
}
\usage{

see.cor.range.tck(sd = 0.5)
}

\arguments{

  \item{sd}{
Amount of noise added to linear association.  Residuals around line pulled from a normal distribution centered at zero with this standard deviation.  
}
}
\references{
Based on a figure from \url{http://en.wikipedia.org/wiki/Correlation_and_dependence}
}
\author{
Ken Aho
}
\keyword{graphs}

