\name{see.exppower.tck}
\alias{see.exppower.tck}

\title{
Visualize exponential power functions
}
\description{
Visualize exponential power functions, including a Gaussian distribution.
}
\usage{

see.exppower.tck()
}

\details{
The normal distribution and Gaussian distribution are based on an exponential power function:
\deqn{f(x)=exp(-|x|^{m})}
Letting \eqn{m = 2} results in a Gaussian distribution.  Standardizing this so that the area under the curve = 1 results in the standard normal distribution. 
}

\author{
Ken Aho
}
\seealso{
\code{\link{demos}}
}

\keyword{graphs}


