\name{see.HW}
\alias{see.HW}
\alias{see.HW.tck}

\title{
Visualize the Hardy Weinberg equlibrium
}
\description{
Allows interactive depiction of the Hardy Weinberg equilibrium.
}
\usage{
see.HW(parg)
see.HW.tck()
}

\arguments{

  \item{parg}{Proportion of the allele \emph{p} in the population, i.e. a number between 0 and 1.

}

}
\details{
Solves and depicts the Hardy Weinberg equilibrium, i.e:
\deqn{pp + 2pq + qq = 1}
}
\author{Ken Aho
}
\keyword{grpahs}

