\name{see.logic}
\alias{see.logic}
\title{
Interactive worksheet for logical and fallacious arguments
}
\description{
It is vital that scientists understand what logical and fallacious arguments are.  This worksheet provides a predagogical tool for logic.
}
\usage{
see.logic()
}
\references{
Salmon, W (1963) \emph{Logic}.  Prentice-Hall
}
\author{
Ken Aho
}
\seealso{
\code{\link{book.menu}}
}
\examples{
\dontrun{
see.logic()
}
}
\keyword{graphs}

