\name{see.M}
\alias{see.M}
\title{Visualization of the M-estimation function
}
\description{
The function provides interactive visualization of robust \emph{M} estimation of location.  
}
\usage{
see.M()
}
\details{
The value \emph{c} = 1.28 gives 95 percent efficiency of the mean given normality.  The sample median and mean can be considered special cases of \emph{M}-estimators.  The value \emph{c} = 0 provides the sample median, while the value, \eqn{c = \infty} gives the sample mean.
}

\references{
Wilcox, R. R. (2005)  \emph{Introduction to Robust Estimation and Hypothesis Testing, Second 
Edition}.  Elsevier, Burlington, MA.
}
\author{
Ken Aho
}
\seealso{
\code{\link{huber.mu}}, \code{\link{huber.NR}}
}
\keyword{graphs}

