\name{see.nlm}
\alias{see.nlm}

\title{
Visualize important non-linear functions
}
\description{
A number of important equation forms require that their parameters be estimated using the non-linear least squares.  Here are six.
}
\usage{
see.nlm()
}

\references{Crawley, M. J.  (2007)  \emph{The R Book}.  Wiley.
}
\author{
Ken Aho
}
\keyword{graphs}
