\name{see.rEffect.tck}
\alias{see.rEffect.tck}
\alias{r.eff}
\title{
Visualize random effects model
}
\description{
An experiment to ascertain the effect of two randomly selected brands of soil fertilizer 
on wheat yield.  In the upper figure two brands of fertilizer (1 and 2) 
are randomly chosen from a population of potential choices.  The mean yields produced by the 
population of fertilizers E(Y|Ai) are normally distributed.  That is, it is possible to 
select a factor level that will result in very small average yields, or one that will 
result in large average yields, but it is more likely that a chosen factor level will 
produce some intermediate average effect.  We proceed with the experiment by assigning two 
experimental units (two wheat fields) to each fertilizer.  We assume that the yield of 
fields is normally distributed for each fertilizer, and furthermore that the factor 
levels are homoscedastic. We weigh our evidence against the H0 of a non-zero population variance by estimating the 
variability among factor levels.  The more that yield varies with respect to nutrient treatments the more evidence we will have against H0.  
}
\usage{
see.rEffect.tck()
}

\author{Ken Aho
}
\keyword{graphs}


