\name{see.regression}
\alias{see.regression}
\alias{see.adddel}
\title{
Interactive visualization of least squares regression.  
}
\description{
Scatterplot points can be moved with \code{see.regression}, while points can be added and deleted with \code{see.adddel}. The function \code{see.regression} is an appropriation from \pkg{tcltk} demos, with a few bells and whistles added.
}
\usage{
see.regression()
see.adddel()
}
\author{the R Development Core Team for \code{see.regression}, Ken Aho for \code{see.adddel}. 
}
\keyword{graphs}


