\name{see.regression.tck}
\alias{see.regression.tck}

\title{
Demonstration of regression mechanics
}
\description{
Population and sample regression lines are interactively depicted.  The same random observations generated by the true error distribution is used for both models. 
}
\usage{

see.regression.tck()
}
\author{
Ken Aho
}
\keyword{graphs}

