\name{see.roc.tck}
\alias{see.roc.tck}

\title{
Interactive depiction of ROC curves
}
\description{
Sliders allow users to change distinctness of dichotomous classes (success and failure).  
This will affect the ROC curve.  One can also change the criteria defining what constitutes a success.  
While this will not change the ROC curve (which compares true positive and false negative rates at all possible success cutoff), 
it The will change empirical rates of true positives, true negatives, false positives, and false negatives given the defined cutoff.  
}
\usage{
 see.roc.tck()
}

\author{
Ken Aho, inspired by a graphical demo at 
\url{http://www.anaesthetist.com/mnm/stats/roc/Findex.htm} 
}
\keyword{graphs}


