\name{see.smooth.tck}
\alias{see.smooth.tck}
\title{
Interactive smoother demonstrations
}
\description{
LOWESS, kernel, and spline smoothers are depicted, using \pkg{tcltk} widgets.}
\usage{
see.smooth.tck()
}

\author{
Ken Aho, appropriated ideas from demo in library \pkg{tcltk}.
}
\seealso{
\code{\link{loess}}, \code{\link{ksmooth}}, \code{\link{smooth.spline}}
}
\keyword{graphs}


