\name{see.typeI_II}
\alias{see.typeI_II}

\title{
Interactive depiction of type I and II error
}
\description{
The function provides a \pkg{tcltk} GUI illustrating type I, type II error, and power.
}
\usage{
see.typeI_II()
}
\author{Ken Aho
}
\seealso{
\code{\link{see.power}},\code{\link{power.z.test}}
}
\keyword{graphs}


