\name{selftest.se.tck1}
\alias{selftest.conf.tck1}
\alias{selftest.se.tck1}
\alias{selftest.se.tck2}
\alias{selftest.pdfs.tck}
\alias{selftest.prob.tck}
\alias{selftest.sampd.tck1}
\alias{selftest.stats.tck}
\alias{selftest.stats.tck1}
\alias{selftest.H0.tck1}
\alias{selftest.H0.tck2}
\alias{selftest.corr.tck1}
\alias{selftest.regdiag.tck1}
\alias{selftest.regdiag.tck2}
\alias{selftest.regdiag.tck3}
\alias{selftest.regapproaches.tck1}
\alias{selftest.regapproaches.tck2}
\alias{selftest.regapproaches.tck3}
\alias{selftest.regapproaches.tck4}
\alias{selftest.ANOVAsiminf.tck1}
\alias{selftest.ANOVAmixed.tck1}
\alias{selftest.typeIISS.tck1}
\alias{selftest.ML_OLS.tck}

\title{
Interactive self-testing statistical questions
}                                            
\description{
These functions provide interactive multiple-choice questions. 
}
\usage{

selftest.se.tck1()
selftest.se.tck2()
selftest.conf.tck1()
selftest.pdfs.tck()
selftest.prob.tck()
selftest.sampd.tck1()
selftest.stats.tck()
selftest.stats.tck1()


}
\author{
Ken Aho
}
