\name{shad}
\alias{shad}
\docType{data}
\title{
American gizzard shad data
}
\description{
Hollander and Wolfe (1999) describe young of year lengths at four sites for American gizzard shad, \emph{Dorosoma cepedianum}, a fish of the herring family. 
}
\usage{data(shad)}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{length}}{Fish length in cm}
    \item{\code{site}}{a factor with levels \code{I} \code{II} \code{III} \code{IV}}
  }
}

\source{
Hollander, M., and  D. A. Wolfe  (1999) \emph{Nonparametric Statistical Methods}. New York: John Wiley & Sons. 
}
\keyword{datasets}

