\name{shade.norm.tck}
\alias{shade.bin.tck}
\alias{shade.chi.tck}
\alias{shade.F.tck}
\alias{shade.norm.tck}
\alias{shade.poi.tck}
\alias{shade.t.tck}

\title{GUI display of probability 
}
\description{
Provides \pkg{tcltk} GUIs to manage \pkg{asbio} \code{\link{shade}} functions.
}
\usage{

shade.bin.tck()

shade.chi.tck()

shade.F.tck()

shade.norm.tck()

shade.poi.tck()

shade.t.tck()
}

\author{Ken Aho
}

\seealso{\code{\link{shade}}}
\keyword{graphs}







