\name{snore}
\alias{snore}
\docType{data}
\title{
Snoring and heart disease contingency data 
}
\description{
Norton and Dunn (1985) compiled data from four family practice clinics in Toronto to quantify the association between snoring and heart disease for 2484 subjects.  
}
\usage{data(snore)}
\format{
  A data frame with 2484 observations on the following 3 variables.
  \describe{
    \item{\code{snoring}}{A factor with levels \code{every.night} \code{nearly.ever.night} \code{never} \code{occasional}}
    \item{\code{ord.snoring}}{Agresti (21012) transformed the explanatory levels to ordinal values in his analysis of this data.}
    \item{\code{disease}}{Presence/absence of heart disease}
  }
}
\source{
Norton, P. G., and E. V. Dunn  (1985)  Snoring as a risk factor for disease: an 
	epidemiological survey.  \emph{British Medical Journal} 291: 630-632. 
}
\keyword{datasets}

