\name{stan.error}
\alias{stan.error}
\alias{stan.error.sq}
\title{Variance and standard error estimators for the sampling distribution of the sample mean}
\description{Estimator for the variance for the sampling distribution of the sample mean, i.e. \eqn{S^2/n}, and the standard deviation of the sampling distribution, i.e. \eqn{S/\sqrt(n)}.
}
\usage{
stan.error(x)
stan.error.sq(x)
}
\arguments{
\item{x}{A vector of quantitative data.}}
\author{Ken Aho}
\seealso{\code{\link{sd}}}
\keyword{univar}
