\name{starkey}
\alias{starkey}
\docType{data}
\title{DEM data from the Starkey experimental forest in NE Oregon}
\description{
UTM northing and easting data along with 18 other environmental variables describing the
Starkey experimental forest. 
}
\usage{data(starkey)}
%\source{
%\url{https://www.fs.fed.us/pnw/starkey/data/tables/Starkey_OR_Main_Habitat_1993-1996_Data.txt
%}}
\keyword{datasets}
