\name{veneer}
\alias{veneer}
\docType{data}
\title{
Veneer data from Littell et al. (2002)
}
\description{
Four examples of each of five brands of a synthetic wool veneer material were subjected to a friction test, and a measure of wear was determined for each experimental unit. 
}
\usage{data(veneer)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{wear}}{a numeric vector}
    \item{\code{brand}}{a factor with levels \code{ACME} \code{AJAX} \code{CHAMP} \code{TUFFY} \code{XTRA}}
  }
}
\source{
Littell, R. C., Stroup, W. W., and R. J. Fruend (2002)  \emph{SAS for Linear Models}.  John 
Wiley and Associates. }
\keyword{datasets}
