\name{vs}
\alias{vs}
\alias{varespec}
\docType{data}
\title{
Scandinavian site by species community matrix
}
\description{
Scandinavian, lichen, bryophyte, and vascular plant data from Vare et al. (1995).
}
\usage{data(vs)}
\format{
  A data frame with 24 observations (sites) on 44 variables (species).
}
\details{
Lifted from dataframe \code{\link[vegan]{varespec}}.
}
\source{
\code{\link[vegan]{varespec}}
}
\references{
Vare, H., Ohtonen, R., and J. Oksanen (1995) Effects of reindeer grazing on understory vegetation in dry Pinus sylvestris forests. \emph{Journal of Vegetation Science} 6: 523-530.
}
\keyword{datasets}
