\name{wash.rich}
\alias{wash.rich}
\docType{data}
\title{
Species richness and environmental variables from Mt Washburn
}
\description{
Aho and Weaver (2010) examined the effect of environmental characteristics on alpine vascular plant species richness on Mount Washburn (3124m) a volcanic peak in north-central Yellowstone National Park.
}
\usage{data(wash.rich)}
\format{
  A data frame with 40 observations on the following 7 variables.
  \describe{
    \item{\code{site}}{Site identifier.}
    \item{\code{Y}}{Species richness.}
    \item{\code{X1}}{Percent Kjeldahl (total) soil N.}
    \item{\code{X2}}{Slope in degrees.}
    \item{\code{X3}}{Aspect in degrees from true north.}
    \item{\code{X4}}{Percent cover of surface rock.}
    \item{\code{X5}}{Soil pH.}
  }
}
\source{
Aho, K., and T. Weaver (2010)  Ecology of alpine nodes: environments, communities, 
and ecosystem evolution (Mount Washburn; Yellowstone Natl. Park).  \emph{Arctic, Antarctic, and Alpine Research}.  40(2): 139-151. 
}
\keyword{datasets}

