\name{world.co2}
\alias{world.co2}
\docType{data}
\title{World CO2 levels, by country, from 1980 to 2006}
\description{
The US department of energy has compiled data since 1980 detailing total carbon dioxide emissions from the consumption and flaring of fossil fuels (in millions of metric tons of carbon dioxide).  Data can be broken down by country.
}
\usage{data(world.co2)}
\format{
The dataframe contains 16 columns
\describe{
\item{\code{Year}}{The year of CO\eqn{_2} measure (1980-2006)}.   
\item{\code{Afghanistan}}{CO\eqn{_2} emissions in Afghanistan from 1980-2006 (1x\eqn{10^6} metric tons).}         
\item{\code{Belgium}}{CO\eqn{_2} emissions in Belgium...}        
\item{\code{Brazil}}{CO\eqn{_2} emissions in Brazil...}
\item{\code{Canada}}{CO\eqn{_2} emissions in Canada...}
\item{\code{China}}{CO\eqn{_2} emissions in China...}
\item{\code{Finland}}{CO\eqn{_2} emissions in Finland...}
\item{\code{Ghana}}{CO\eqn{_2} emissions in Ghana...}
\item{\code{Italy}}{CO\eqn{_2} emissions in Italy...}        
\item{\code{Japan}}{CO\eqn{_2} emissions in Japan...}
\item{\code{Kenya}}{CO\eqn{_2} emissions in Kenya...}
\item{\code{Mexico}}{CO\eqn{_2} emissions in Mexico...}        
\item{\code{Saudi.Arabia}}{CO\eqn{_2} emissions in Saudi Arabia...}
\item{\code{United.Arab.Emirates}}{CO\eqn{_2} emissions in the United Arab Emirates..}
\item{\code{United.States}}{CO\eqn{_2} emissions in United States...}
\item{\code{World.Total}}{CO\eqn{_2} emissions totals for the world...}
  }
}
\source{US Department of energy; \url{www.eia.doe.gov/emeu/iea/}}
\keyword{datasets}
