\name{world.pop}
\alias{world.pop}
\docType{data}
\title{Population levels in various countries since 1980}
\description{
Population levels of 13 countries since 1980.  Population numbers are rounded to the nearest 100,000.
}
\usage{data(world.pop)}
\format{
The dataframe contains 14 columns
\describe{
\item{Year}{The year of population measurements (1980-2006)}   
\item{Afghanistan}{Population in Afghanistan from 1980-2006, rounded to the nearest 100,000.}         
\item{Brazil}{Population in Brazil...}
\item{Canada}{Population in Canada...}
\item{China}{Population in China...}
\item{Finland}{Population in Finland...}
\item{Italy}{Population in Italy...}        
\item{Japan}{Population in Japan...}
\item{Kenya}{Population in Kenya...}
\item{Mexico}{Population in Mexico...}        
\item{Saudi.Arabia}{Population in Saudi Arabia...}
\item{United.Arab.Emirates}{Population in the United Arab Emirates...}
\item{United.States}{Population in United States...}
\item{World.Total}{Population totals for the world...}
  }
}

\source{
US census bureau; \url{http://www.census.gov/ipc/www/idb/idbsprd.html}
}
\keyword{datasets}
