\name{agrostis}
\alias{agrostis}
\docType{data}
\title{
Agrostis variabilis cover measurements
}
\description{
Percent cover of the grass \emph{Agrostis variablis} at 25 alpine snowbank sites in the Absaroka-Beartooth Mountains.
}

\usage{data(agrostis)}
\format{
  A data frame with 25 observations on the following 2 variables.
  \describe{
    \item{site}{Site number}
    \item{cover}{Percent cover}
  }
}
\source{
Aho, K.  (2006)  \emph{Alpine and Cliff Ecosystems in the North-Central Rocky Mountains}.  PhD dissertation, Montana State University, Bozeman, MT.
}
\keyword{datasets}

