\name{demos}
\alias{demos}
\title{Pulldown menu for 'asbio' interactive graphical functions.
}
\description{
Provides a pulldown GUI menu to allow access to \pkg{asbio} statistical and biological graphical demos, e.g. \code{\link{anm.ci}}, \code{\link{samp.dist}}, \code{\link{loglik.plot}}, etc.  The function currently provides links to 93 distinct GUIs (many of these are slaves to other GUIs), which in turn provide distinct graphical demonstrations. The GUIs work best with an SDI R interface.   
}
\usage{

demos()
}


\author{Ken Aho}

\seealso{
\code{\link{anm.coin}},  \code{\link{anm.ci}},  \code{\link{anm.die}},  \code{\link{anm.exp.growth}},  \code{\link{anm.geo.growth}},  \code{\link{anm.log.growth}}, \code{\link{loglik.plot}},  \code{\link{anm.LVcomp}},  \code{\link{anm.LVexp}},  \code{\link{anm.ls}},  \code{\link{anm.transM}},  \code{\link{samp.dist}}, \code{\link{see.HW}},  \code{\link{see.logic}},  \code{\link{see.M}},  \code{\link{see.norm.tck}},  \code{\link{see.power}},  \code{\link{see.regression}},  \code{\link{selftest.se.tck1}},  \code{\link{shade.norm}},  \code{\link{see.typeI_II}}, \code{\link{Venn}}} 
\examples{
\dontrun{
demos()
}
}
\keyword{graphs}

