\name{illusions}
\alias{illusions}
\title{
Visual illusions illustrating human preception errors.
}
\description{
In development, currently displays two illusions.
}
\usage{

illusions(ill.no = 1)
}

\arguments{

  \item{ill.no}{Numeric describing which illusion number to view.

}
}

\author{
Ken Aho
}

\keyword{graphs}
