\name{illusions}
\alias{illusions}
\title{
Visual illusions illustrating human preception errors.
}
\description{
In development, currently displays three illusions.  Illusion 3 is from Yihui Xie's package \pkg{animation}.
}
\usage{

illusions(ill.no = 1)
}

\arguments{

  \item{ill.no}{Numeric describing which illusion number to view.

}
}

\author{
Ken Aho
}
\examples{
illusions(2)
illusions(3)
}
\keyword{graphs}
