\name{samp.dist.mech}
\alias{samp.dist.mech}

\title{
Animated representation of sampling distribution basics
}
\description{
Goats are sampled 10 at a time and weighed, a mean weight is calculated from these measures and added to collection of mean weights in the form of a histogram. 
}
\usage{

samp.dist.mech(rep, int = 0.05)
}
\arguments{

  \item{rep}{
Number of samples.  Should not greatly exceed 100.
}
  \item{int}{The time interval for animation (in seconds).  Smaller intervals speed up animation}
}

\author{
Ken Aho
}
\note{
Nice goat image from http://all-free-download.com 
}

\keyword{graphs}

