\name{see.anova.tck}
\alias{see.anova.tck}

\title{
Interactive depiction of the ANOVA mechanism
}
\description{
Slider control of the means and (constant) variability of three factor level populations.  An ANOVA is run based on a random sample of these populations.
}
\usage{

see.anova.tck()
}


\author{
Ken Aho
}
\keyword{graphs}
