\name{see.nlm}
\alias{see.nlm}

\title{
Visualize important non-linear equations
}
\description{
A number of important equation forms have their paramters estimated using the non-linear least squares.  Here are six.
}
\usage{
see.nlm()
}

\references{Crawley, M. J.  (2007)  \emph{The R book}.  Wiley.
}
\author{
Ken Aho
}
\keyword{grpahs}
