\name{see.norm.tck}
\alias{see.norm.tck}
\title{Visualize the normal distribution}
\description{
An interactive GUI for visualizing how the normal distribution changes with changing values of \eqn{\mu} and \eqn{\sigma}.  Lifted almost verbatim from a clever function from Greg Snow's package \pkg{TeachingDemos}. 
}
\usage{
see.norm.tck()
}
\author{Greg Snow \email{greg.snow@imail.org}.  Wrapper for Snow's \code{vis.normal} written by Ken Aho.
}
\examples{
\dontrun{
see.norm.tck()
}
}
\keyword{graphs}
