\name{selftest.se.tck1}
\alias{selftest.ANVOVA_design.tck}
\alias{selftest.ANOVAsiminf.tck1}
\alias{selftest.ANOVAmixed.tck1}
\alias{selftest.corr.tck1}
\alias{selftest.conf.tck1}
\alias{selftest.bayes4.tck}
\alias{selftest.bayes5.tck}
\alias{selftest.H0.tck}
\alias{selftest.jack.tck}
\alias{selftest.ML_OLS.tck}
\alias{selftest.nonparametric6.tck}
\alias{selftest.pdfs.tck}
\alias{selftest.prob.tck}
\alias{selftest.regchar.tck1}
\alias{selftest.regdiag.tck1}
\alias{selftest.regdiag.tck2}
\alias{selftest.regdiag.tck3}
\alias{selftest.regapproaches.tck1}
\alias{selftest.regapproaches.tck2}
\alias{selftest.regapproaches.tck3}
\alias{selftest.regapproaches.tck4}
\alias{selftest.sampd.tck1}
\alias{selftest.se.tck1}
\alias{selftest.stats.tck}
\alias{selftest.stats.tck1}
\alias{selftest.ttest.tck}
\alias{selftest.codettest.tck}
\alias{selftest.typeIISS.tck1}

\title{
Interactive self-testing questions
}                                            
\description{
These functions provide interactive multiple-choice questions. 
}
\usage{
selftest.se.tck1()
}
\author{
Ken Aho
}
