\name{selftest.se.tck1}
\alias{selftest.se.tck1}
\alias{selftest.se.tck2}
\alias{selftest.pdfs.tck1}
\alias{selftest.pdfs.tck2}
\title{
Interactive self-testing statistical questions
}                                            
\description{
These functions provide interactive multiple-choice questions that can be used to emphasize pedagogical ideas. 
}
\usage{

selftest.se.tck1()
selftest.se.tck2()
selftest.pdfs.tck1()
selftest.pdfs.tck2()

}
\author{
Ken Aho
}
