\name{selftest.se.tck1}
\alias{selftest.conf.tck1}
\alias{selftest.conf.tck2}
\alias{selftest.se.tck1}
\alias{selftest.se.tck2}
\alias{selftest.pdfs.tck1}
\alias{selftest.pdfs.tck2}
\alias{selftest.pdfs.tck3}
\alias{selftest.prob.tck1}
\alias{selftest.prob.tck2}
\alias{selftest.prob.tck3}
\alias{selftest.sampd.tck1}
\alias{selftest.sampd.tck2}
\alias{selftest.stats.tck1}
\alias{selftest.stats.tck2}
\alias{selftest.H0.tck1}
\alias{selftest.H0.tck2}
\alias{selftest.corr.tck1}
\alias{selftest.regdiag.tck1}
\alias{selftest.regdiag.tck2}
\alias{selftest.regdiag.tck3}
\alias{selftest.regapproaches.tck1}
\alias{selftest.regapproaches.tck2}
\alias{selftest.regapproaches.tck3}
\alias{selftest.regapproaches.tck4}
\alias{selftest.ANOVAsiminf.tck1}
\alias{selftest.ANOVAmixed.tck1}
\alias{selftest.typeIISS.tck1}

\title{
Interactive self-testing statistical questions
}                                            
\description{
These functions provide interactive multiple-choice questions. 
}
\usage{

selftest.se.tck1()
selftest.se.tck2()
selftest.conf.tck1()
selftest.conf.tck2()
selftest.pdfs.tck1()
selftest.pdfs.tck2()
selftest.pdfs.tck3()
selftest.prob.tck1()
selftest.prob.tck2()
selftest.prob.tck3()
selftest.sampd.tck1()
selftest.sampd.tck2()
selftest.stats.tck1()
selftest.stats.tck2()

}
\author{
Ken Aho
}
