\name{trag}
\alias{trag}
\docType{data}
\title{
Salsify height dataset
}
\description{
Heights of slasify \emph{Tragapogon dubius} at the Barton Road long term experiental site in Pocatello Idaho. 
}

\usage{data(trag)}


\format{
  A data frame with 20 observations on the following variable.
  \describe{
    \item{\code{height}}{\emph{T. dubius} plant height in cm}
  }
}
\keyword{datasets}

