\name{webs}
\alias{webs}
\docType{data}
\title{Spider web length data}
\description{
Gosline et al. (1984) applied heat to strands of spider web to determine whether the structure underlying webs was rubber-like. Data are estimated from a scatterplot in Gosline et al.. 
}
\usage{data(webs)}
\format{
The dataframe contains 4 columns
\describe{
\item{obs}{Observation number.}   
\item{relative length}{Relative strand strand length after heat application.}         
\item{temp.C}{Temp in degrees celsius.}        
\item{residuals}{Residuals from the linear model length~temp.C.}
  }
}
\source{Gosline, J. M., Denny, M. W. and Demont, M. E. (1984). Spider silk as rubber.
\emph{Nature} 309, 551-552.}
\keyword{datasets}
