\name{world.pop}
\alias{world.pop}
\docType{data}
\title{Population levels in various countries from 1980-2006}
\description{
Population levels of 13 countries from 1980-2006.  Population numbers are rounded to the nearest 100,000.  More up-to-date data can be found in this package at \code{\link{world.emissions}}
}
\usage{data(world.pop)}
\format{
The dataframe contains 14 columns
\describe{
\item{\code{Year}}{The year of population measurements (1980-2006)}   
\item{\code{Afghanistan}}{Population in Afghanistan from 1980-2006, rounded to the nearest 100,000.}         
\item{\code{Brazil}}{Population in Brazil...}
\item{\code{Canada}}{Population in Canada...}
\item{\code{China}}{Population in China...}
\item{\code{Finland}}{Population in Finland...}
\item{\code{Italy}}{Population in Italy...}        
\item{\code{Japan}}{Population in Japan...}
\item{\code{Kenya}}{Population in Kenya...}
\item{\code{Mexico}}{Population in Mexico...}        
\item{\code{Saudi.Arabia}}{Population in Saudi Arabia...}
\item{\code{United.Arab.Emirates}}{Population in the United Arab Emirates...}
\item{\code{United.States}}{Population in United States...}
\item{\code{World.Total}}{Population totals for the world...}
  }
}

\source{
US census bureau: \url{https://www.census.gov/programs-surveys/international-programs/about/idb.html}
}
\keyword{datasets}
